fixed complex cohort targeting a refactored the code to make the files a bit less unwieldy
