+++ rust/feature-flags/src/flags/flag_operations.rs
@@ -1,6 +1,6 @@
 use crate::api::errors::FlagError;
 use crate::client::database::Client as DatabaseClient;
-use crate::cohort::cohort_models::CohortId;
+use crate::cohorts::cohort_models::CohortId;
 use crate::flags::flag_models::*;
 use crate::properties::property_models::PropertyFilter;
 use common_redis::Client as RedisClient;
