+++ rust/feature-flags/src/flags/mod.rs
@@ -1,6 +1,8 @@
 pub mod flag_analytics;
+pub mod flag_group_type_mapping;
 pub mod flag_match_reason;
 pub mod flag_matching;
+pub mod flag_matching_utils;
 pub mod flag_models;
 pub mod flag_operations;
 pub mod flag_request;
