+++ rust/feature-flags/src/router.rs
@@ -20,7 +20,7 @@ use tower_http::{
 use crate::{
     api::{endpoint, test_endpoint},
     client::database::Client as DatabaseClient,
-    cohort::cohort_cache_manager::CohortCacheManager,
+    cohorts::cohort_cache_manager::CohortCacheManager,
     config::{Config, TeamIdsToTrack},
     metrics::utils::team_id_label_filter,
 };
