+++ rust/feature-flags/src/utils/test_utils.rs
@@ -1,16 +1,16 @@
 use crate::{
     client::database::{get_pool, Client, CustomDatabaseError},
-    cohort::cohort_models::{Cohort, CohortId},
+    cohorts::cohort_models::{Cohort, CohortId},
     config::{Config, DEFAULT_TEST_CONFIG},
-    flags::{
-        flag_matching::PersonId,
-        flag_models::{FeatureFlag, FeatureFlagRow, TEAM_FLAGS_CACHE_PREFIX},
+    flags::flag_models::{
+        FeatureFlag, FeatureFlagRow, FlagFilters, FlagGroupType, TEAM_FLAGS_CACHE_PREFIX,
     },
     team::team_models::{Team, TEAM_TOKEN_CACHE_PREFIX},
 };
 use anyhow::Error;
 use axum::async_trait;
 use common_redis::{Client as RedisClientTrait, RedisClient};
+use common_types::{PersonId, TeamId};
 use rand::{distributions::Alphanumeric, Rng};
 use serde_json::{json, Value};
 use sqlx::{pool::PoolConnection, postgres::PgRow, Error as SqlxError, Postgres, Row};
@@ -506,4 +506,39 @@ pub async fn create_group_in_pg(
         group_key: group_key.to_string(),
         group_properties,
     })
+}
+
+#[allow(clippy::too_many_arguments)]
+pub fn create_test_flag(
+    id: Option<i32>,
+    team_id: Option<TeamId>,
+    name: Option<String>,
+    key: Option<String>,
+    filters: Option<FlagFilters>,
+    deleted: Option<bool>,
+    active: Option<bool>,
+    ensure_experience_continuity: Option<bool>,
+) -> FeatureFlag {
+    FeatureFlag {
+        id: id.unwrap_or(1),
+        team_id: team_id.unwrap_or(1),
+        name: name.or(Some("Test Flag".to_string())),
+        key: key.unwrap_or_else(|| "test_flag".to_string()),
+        filters: filters.unwrap_or_else(|| FlagFilters {
+            groups: vec![FlagGroupType {
+                properties: Some(vec![]),
+                rollout_percentage: Some(100.0),
+                variant: None,
+            }],
+            multivariate: None,
+            aggregation_group_type_index: None,
+            payloads: None,
+            super_groups: None,
+            holdout_groups: None,
+        }),
+        deleted: deleted.unwrap_or(false),
+        active: active.unwrap_or(true),
+        ensure_experience_continuity: ensure_experience_continuity.unwrap_or(false),
+        version: Some(1),
+    }
 }
